+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -16,6 +16,7 @@ group: Deployment & Configs
 - Added 'Add Toposolid Sub-Division' component.
 - Added 'Component References' component.
 - Added 'Component Reference Plane' component.
+- Added 'Element Annotations' component.
 
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}